+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Add Proton Sarek for legacy graphics cards.
 - Add hibernation support for NVIDIA graphics cards.
 - Change more first-time setup tasks to be included in the productivity section instead of separate sections.
+- Add the Homebrew package manager.
 
 ## 4.1.0 - The Dual-Boot Update